+++ src/command/commands/opponent.ts
@@ -27,7 +27,7 @@ const command: Command = {
 
     if (!ctx.tags.owner && !king.startsWith('secret_')) {
       void services.twitchService.ownerRunCommand(
-        `!alert ${ctx.tags['display-name']} changed my opponent's king to ${king}`
+        `!alert {${ctx.tags['display-name']}} changed my opponent's king to {${king}}`
       );
     }
   }